requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
Sphinx==1.7.0

pytest==3.4.1
pytest-runner==2.11.1
pytest-watch==4.1.0
mock>=2.0.0
mockredispy>=2.9.3
